+++ src/H5Tpublic.h
@@ -293,8 +293,42 @@ H5_DLLVAR hid_t H5T_IEEE_F64LE_g;
  * 16-bit little-endian bfloat16 floating-point numbers
  */
 #define H5T_FLOAT_BFLOAT16LE (H5OPEN H5T_FLOAT_BFLOAT16LE_g)
+/**
+ * \ingroup PDTALTFLOAT
+ * 8-bit FP8 E4M3 (4 exponent bits, 3 mantissa bits) floating-point numbers
+ *
+ * \parblock
+ * \attention Implicit datatype conversion should currently be avoided when using
+ *            this datatype for I/O operations. Incomplete handling of non-IEEE
+ *            floating-point formats in HDF5 can cause certain FP8 E4M3 values
+ *            to be improperly converted to Infinities or NaN values. If possible,
+ *            an application should perform I/O with this datatype using an
+ *            in-memory type that matches the FP8 E4M3 format and perform explicit
+ *            data conversion outside of HDF5, if necessary. Otherwise, read/written
+ *            values should be verified to be correct.
+ * \endparblock
+ */
+#define H5T_FLOAT_F8E4M3 (H5OPEN H5T_FLOAT_F8E4M3_g)
+/**
+ * \ingroup PDTALTFLOAT
+ * 8-bit FP8 E5M2 (5 exponent bits, 2 mantissa bits) floating-point numbers
+ *
+ * \parblock
+ * \attention Implicit datatype conversion should currently be avoided when using
+ *            this datatype for I/O operations. Incomplete handling of non-IEEE
+ *            floating-point formats in HDF5 can cause certain FP8 E5M2 values
+ *            to be improperly converted to Infinities or NaN values. If possible,
+ *            an application should perform I/O with this datatype using an
+ *            in-memory type that matches the FP8 E5M2 format and perform explicit
+ *            data conversion outside of HDF5, if necessary. Otherwise, read/written
+ *            values should be verified to be correct.
+ * \endparblock
+ */
+#define H5T_FLOAT_F8E5M2 (H5OPEN H5T_FLOAT_F8E5M2_g)
 H5_DLLVAR hid_t H5T_FLOAT_BFLOAT16BE_g;
 H5_DLLVAR hid_t H5T_FLOAT_BFLOAT16LE_g;
+H5_DLLVAR hid_t H5T_FLOAT_F8E4M3_g;
+H5_DLLVAR hid_t H5T_FLOAT_F8E5M2_g;
 
 /*
  * Complex number types made up of IEEE floating point types